Split up composiite dummy message class into individual items
