+++ ruby/spec/spec_helper.rb
@@ -1,5 +1,8 @@
 require 'rspec'
 
 require 'cucumber/messages'
+require 'cucumber/messages/message'
 
-require_relative 'support/dummy_messages'
+require_relative 'support/comprehensive_message'
+require_relative 'support/enum_message'
+require_relative 'support/simple_message'
