+++ ruby/spec/support/comprehensive_message.rb
@@ -1,22 +1,7 @@
-require 'cucumber/messages/message'
+require_relative 'simple_message'
 
 module Cucumber
   module Messages
-    class SimpleMessage < Message
-      attr_reader :is_nil, :is_string, :is_array, :is_number
-
-      def initialize(is_nil: nil, is_string: '', is_array: [], is_number: 0)
-        @is_nil = is_nil
-        @is_string = is_string
-        @is_array = is_array
-        @is_number = is_number
-      end
-    end
-
-    class EnumMessage
-      ENUM = 'an enum'
-    end
-
     class ComprehensiveMessage < Message
       attr_reader :simple_message, :message_array, :is_enum
