+++ ruby/spec/support/enum_message.rb
@@ -1,0 +1,7 @@
+module Cucumber
+  module Messages
+    class EnumMessage
+      ENUM = 'an enum'
+    end
+  end
+end
